update WalletTransactionResponseDto and WalletTransaction entity

+++ src/main/java/com/spareparts/spareparts_backend/dto/WalletTransactionResponseDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,20 +14,12 @@ import java.time.LocalDateTime;
 @Builder
 public class WalletTransactionResponseDto {
     private Integer transactionId;
-
     private Integer customerId;
-
-    private double amount;                 // + credit / - debit
-
+    private BigDecimal amount;                 // + credit / - debit
     private String transactionType;        // TOP_UP, ORDER_PAYMENT, REFUND
-
     private String referenceType;           // ORDER, REFUND, MANUAL
-
     private Integer referenceId;            // orderId / refundId
-
-    private double balanceAfter;            // wallet balance after txn
-
+    private BigDecimal balanceAfter;            // wallet balance after txn
     private String description;
-
     private LocalDateTime createdAt;
 }
